new module is created
